Refresh containers after 30s to work around bug
It seems that containers started at the same time as Dotege
(e.g. by docker-compose) aren't included in the initial list or
the overlapping event stream, somehow. Presumably the docker
daemon is still processing the starts on another thread and
there's a bit of a race condition?

For now, schedule a refresh for 30 seconds after startup to
try and mitigate the problem.