Refactor templates a little.
This reduces the cross-file references going on - containers
don't need to know the context that the template generator
expects; the template generator doesn't need to know about
the config format.

Remove the overall template generator struct as well, as it
was just holding a slice of templates. Might as well just
expose that.